chore(remake vocabulary): add auto play
