+++ src/modules/exercise/constants.ts
@@ -4,12 +4,14 @@ export const EXERCISE_CONFIG_LABELS = {
 	MODE: 'Exercise Mode',
 	CORRECT_RESPONSE: 'Count of total correct response',
 	FORMAT: 'Use a exercise or exercise explanation',
+	AUTO_PLAY: 'Auto play',
 } as const;
 
 export const EXERCISE_CONFIG = {
 	MODE: 'exerciseMode',
 	TOTAL_CORRECT_RESPONSE: 'exerciseCorrectResponse',
 	FORMATE: 'exerciseFormate',
+	AUTO_PLAY: 'autoPlay',
 } as const;
 
 export const MAX_VARIANTS = 8;
