+++ src/modules/vocabularies/VocabularyCard.tsx
@@ -1,23 +1,22 @@
 import { useCallback, useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { ExerciseType, useVocabularyStore } from '@/store/vocabulary-collection';
-import {
-	DEFAULT_DATA_TEST,
-	EXERCISE_FORMATE,
-	EXERCISE_MODE,
-} from '@/store/vocabulary-collection/useVocabularyStore';
 import { useVocabularyListData } from './hooks/useVocabularyListData';
 import { useProgressStore } from '@/store/progress';
 import { WrapperCard } from '@/ui-components/Wrapper-card';
 import { Icon } from '@/ui-components/Icon';
 import { Button } from '@/ui-components/Button';
 import { Result } from '@/ui-components/Result';
-import { ExerciseUI } from '../../shared/components/exercise-content/ExerciseContent';
-import { SelectingUI } from '../../shared/components/selecting-formate-ui/SelectingUI';
 import { PaginationExercise } from './components/pagination/Pagination';
 import { saveVocabularyProgress } from './utils/saveVocabularyProgress';
 import { useBeforeunload } from '@/shared/hooks/use-before-unload/useBeforeunload';
 import styles from './vocabularyCard.module.css';
+import { SelectFormate } from './components/select-formate/SelectFormate';
+import {
+	DEFAULT_DATA_TEST,
+	EXERCISE_FORMATE,
+	EXERCISE_MODE,
+} from '@/store/vocabulary-collection/constants';
 
 export const VocabularyCard = () => {
 	const navigate = useNavigate();
@@ -41,6 +40,7 @@ export const VocabularyCard = () => {
 	const random =
 		randomModeProgress.resolved.every((item) => item.isDone === true) &&
 		randomModeProgress.resolved.length === words.length;
+	const isAutoPlay = useVocabularyStore.use.collectionsExerciseConfig().autoPlay;
 	const getExerciseById = useVocabularyStore.use.getExerciseById();
 	const setExerciseListResponse = useVocabularyStore.use.setExerciseListResponse();
 	const setExerciseListProgress = useVocabularyStore.use.setExerciseListProgress();
@@ -114,21 +114,14 @@ export const VocabularyCard = () => {
 						}
 					/>
 				)}
-				{!testIsDone && task && !isSelectingFormate && (
-					<ExerciseUI
+				{!testIsDone && task && (
+					<SelectFormate
+						isSelectingFormate={isSelectingFormate}
 						key={vocabulariesId}
-						task={task}
-						setIsAutoNavigate={setIsAutoNavigate}
+						task={{ ...task, used: '' }}
+						isAutoPlay={isAutoPlay}
 						setTask={setTask}
-						updateProgress={setModesProgress}
-					/>
-				)}
-				{!testIsDone && task && isSelectingFormate && (
-					<SelectingUI
-						key={vocabulariesId}
-						task={task}
 						setIsAutoNavigate={setIsAutoNavigate}
-						setTask={setTask}
 						updateProgress={setModesProgress}
 					/>
 				)}
