+++ src/shared/constants/vocabulary/collections.ts
@@ -12,32 +12,44 @@ export const DEFAULT_COLLECTIONS_VOCABULARY = [
 		id: '1',
 		title: 'Family',
 		subtitle: 'Family',
-		topic: ['Family'],
+		level: 'A1',
 		category: ['Other'],
-		level: ['A1', 'B2'],
+		topic: ['Family'],
+		learningStyle: 'Selecting/Matching',
+		learnByInterest: 'Books',
+		learnBySkill: 'Reading',
 	},
 	{
 		id: '2',
 		title: 'Parts of the body',
 		subtitle: 'Family',
-		topic: ['Parts of the body'],
+		level: 'B2',
 		category: ['Other', 'Adjectives'],
-		level: ['A1', 'B2', 'C1'],
+		topic: ['Parts of the body'],
+		learningStyle: 'Selecting/Matching',
+		learnByInterest: 'Books',
+		learnBySkill: 'Reading',
 	},
 	{
 		id: '3',
 		title: 'Clothes',
 		subtitle: 'Clothes',
-		topic: ['Clothes'],
+		level: 'C1',
 		category: ['Other', 'Adjectives', 'Verbs'],
-		level: ['A1', 'B2', 'C1', 'C2'],
+		topic: ['Clothes'],
+		learningStyle: 'Selecting/Matching',
+		learnByInterest: 'Books',
+		learnBySkill: 'Reading',
 	},
 	{
 		id: '4',
 		title: 'Describing people',
 		subtitle: 'Describing people',
-		topic: ['Describing people'],
+		level: 'B2',
 		category: ['Adjectives', 'Verbs'],
-		level: ['B1', 'B2'],
+		topic: ['Describing people'],
+		learningStyle: 'Selecting/Matching',
+		learnByInterest: 'Books',
+		learnBySkill: 'Reading',
 	},
 ];
